% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ce.R
\name{analyse_ce}
\alias{analyse_ce}
\title{Analyse the Cost Effectiveness of Interventions}
\usage{
analyse_ce(
  markov_simulations = NULL,
  baseline = 1,
  willingness_to_pay_thresold = 20000,
  type = "base"
)
}
\arguments{
\item{markov_simulations}{A dataframe of markov samples and simulations as produced by \code{markov_simulation_pipeline}. At least two
interventions must be present.}

\item{baseline}{Numeric, the intervetion to consider as the baseline for pairwise comparisons.}

\item{willingness_to_pay_thresold}{Numeric, defaulting to 20,000. This is the thresold at which an intervention
may be considered cost effective in the UK.}

\item{type}{A character string specifying the approach to use to simulate the model. Currently implemented
approaches are "base" with "base" as the default.}
}
\value{
A list of dataframes including: Cost effectiveness measures for each sample, and summarised cost effectiveness
measures across samples.
}
\description{
This function produces cost effectiveness summary measures using the output of \code{markov_simulation_pipeline}
or similar data structures. At least two intervetions must be present.
}
\examples{

sims <-  markov_simulation_pipeline(example_two_state_markov(),
                                    duration = 10, samples = 10)
  
analyse_ce(sims)
  
}
