% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ArmaMarkovLoop}
\alias{ArmaMarkovLoop}
\title{An inner Markov loop implemented using RcppArmadillo}
\usage{
ArmaMarkovLoop(sim, cohort, transition, duration)
}
\arguments{
\item{sim}{Matrix with the same number of rows as the duration of the model and the same number of columns as the number of
states in the model.}

\item{cohort}{A vector containing the initial state.}

\item{transition}{A transition matrix, see \code{example_two_state_markov} for an example of setting this up.}

\item{duration}{Numeric, how many long to run the model for.}
}
\value{
A matrix of the same size as the inputted sim matrix.
}
\description{
An inner Markov loop implemented using RcppArmadillo
}
\examples{

transition <- matrix(rnorm(4), 2, 2)
sim <- matrix(NA, 100, 2)
cohort <- c(1, 0)
duration <- 100

# Reference R implementation
sim_r <- markov_loop(sim, cohort, transition, duration)

# RcppArmadillo implementation
sim_rcppArma <- ArmaMarkovLoop(sim, cohort, transition, duration)

# Check results are within tolerances
all.equal(sim_r, sim_rcppArma)

# Benchmark
library(microbenchmark)
microbenchmark(markov_loop(sim, cohort, transition, duration), 
               ArmaMarkovLoop(sim, cohort, transition, duration), 
               times = 1000)
}
