% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_discounting.R
\name{calc_discounting}
\alias{calc_discounting}
\title{Calculate Discounting Over the Model Run Time.}
\usage{
calc_discounting(discount = 1.035, duration = NULL)
}
\arguments{
\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{duration}{Numeric, how many long to run the model for.}
}
\value{
A vector of length \code{duration}.
}
\description{
Calculate Discounting Over the Model Run Time.
}
\examples{

calc_discounting(1.035, 10)
}
