% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_ce_pipeline.R
\name{markov_ce_pipeline}
\alias{markov_ce_pipeline}
\title{Markov Sampling, Simulation, and Cost Effectiveness Analysis Pipeline}
\usage{
markov_ce_pipeline(
  markov_model = NULL,
  duration = NULL,
  discount = 1.035,
  samples = 1,
  baseline = 1,
  willingness_to_pay_threshold = 20000,
  sample_type = "rcpp",
  sim_type = "armadillo_all",
  debug = FALSE,
  batches = 1,
  batch_fn = NULL,
  ...
)
}
\arguments{
\item{markov_model}{A list of functions that define a markov model across multiple interventions. See \code{example_two_state_markov}
for the correct format.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{samples}{Numeric, defaults to 1. The number of markov model samples to use.}

\item{baseline}{Numeric, the intervention to consider as the baseline for pairwise comparisons.}

\item{willingness_to_pay_threshold}{Numeric, defaulting to 20,000. This is the threshold at which an intervention
may be considered cost effective in the UK.}

\item{sample_type}{A character string specifying the approach to use to sample the model.
Options and defaults inherited from \code{sample_markov}.}

\item{sim_type}{A character string specifying the approach to use to simulate the model.
Options and defaults inherited from \code{simulate_markov}.}

\item{debug}{Logical, defaults to \code{FALSE}. Turns on all debug checks - this may impact runtimes.}

\item{batches}{Numeric, defaults to 1. The number of batches to run simulation/sampling in. When set to
values greater than 1 a \code{batch_fn} must also be supplied. It is likely that the most efficient option will
be to use a batch number that corresponds to the number of cores being utilised.}

\item{batch_fn}{Function, defaults to \code{NULL}. This is the function to be used to parallise across batches. Potential options
include \code{parallel::mclapply} (not Windows) or \code{furrr::future_map} (requires the use of \code{future::plan} outside the function). When
not given the function will default to using no batching.}

\item{...}{Additional options to pass to \code{batch_fn}. For example this may be the \code{mc.cores} argument of \code{parallel::mclapply}.}
}
\value{
A list containing the model samples and simulations and cost effectiveness summary measures.
}
\description{
This functions wraps multiple modular functions and allows an end-to-end cost effectiveness to
be run, including final analysis of the findings.
}
\examples{

markov_ce_pipeline(example_two_state_markov(), duration = 10, samples = 5)
  
}
\seealso{
markov_simulation_pipeline analyse_ce
}
