% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_loop.R
\name{markov_loop}
\alias{markov_loop}
\title{Inner Markov Loop in base R}
\usage{
markov_loop(sim = NULL, cohort, transition = NULL, duration = NULL)
}
\arguments{
\item{sim}{Matrix with the same number of rows as the duration of the model and the same number of columns as the number of
states in the model.}

\item{cohort}{A vector containing the initial state.}

\item{transition}{A transition matrix, see \code{example_two_state_markov} for an example of setting this up.}

\item{duration}{Numeric, how many long to run the model for.}
}
\value{
A matrix of the same size as the inputed sim matrix.
}
\description{
Inner Markov Loop in base R
}
\examples{

 
transition <- matrix(rnorm(4), 2, 2)
sim <- matrix(NA, 10, 2)
cohort <- c(1, 0)
duration <- 10
 
markov_loop(sim, cohort, transition, duration)
}
