% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_arrange.R
\name{matrix_arrange_inner}
\alias{matrix_arrange_inner}
\title{Arrange Vectorised Matrix Samples using R}
\usage{
matrix_arrange_inner(samples)
}
\arguments{
\item{samples}{A list of vectorised matrix samples}
}
\value{
A list of matrices with each matrix representing a single sample.
}
\description{
A convenience function used to arrange vectorised matrix samples into the correct
matrix format for several functions used to specify Markov model. See \code{example_two_state_markov}
for an example use case. Implemented using R.
}
\examples{

matrix_samples <- list(VGAM::rdiric(1:5, c(88, 12)),
                       VGAM::rdiric(1:5, c(8, 92)))

matrix_arrange_inner(matrix_samples)
}
