% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatrixArrange}
\alias{MatrixArrange}
\title{Arrange Vectorised Matrix Samples using Rcpp}
\usage{
MatrixArrange(samples)
}
\arguments{
\item{samples}{A list of vectorised matrix samples}
}
\value{
A list of matrices with each matrix representing a single sample.
}
\description{
A convenience function used to arrange vectorised matrix samples into the correct
matrix format for several functions used to specify Markov model. See \code{example_two_state_markov}
for an example use case. Implemented using Rcpp.
}
\examples{


matrix_samples <- list(VGAM::rdiric(1:5, c(88, 12)),
                      VGAM::rdiric(1:5, c(8, 92)))

# R implementation
samples_r <- matrix_arrange_inner(matrix_samples)


# Rcpp implementation
samples_rcpp <- MatrixArrange(matrix_samples)

all.equal(samples_r, samples_rcpp)

# Benchmark
library(microbenchmark)
microbenchmark(matrix_arrange_inner(matrix_samples), 
               MatrixArrange(matrix_samples), 
               times = 1000)
}
