% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_two_state_markov.R
\name{reference_two_state_markov}
\alias{reference_two_state_markov}
\title{Reference Two State Markov Model}
\usage{
reference_two_state_markov(cycles = NULL, samples = NULL)
}
\arguments{
\item{cycles}{Numeric, the number of cylces (time horizon / cycle length). No default supplied.}

\item{samples}{Numeric, the number of samples to use. No default supplied.}
}
\value{
Output
}
\description{
This is a two state Markov model - modelling smoking cessation used as a baseline and reference.
It has been developed primarily using base R and makes use of a nested array structure. This array is then looped
over using a series of nested for loops with a core loop running a the markov model for each sample and intervention.
Profiling suggestions that this core loop may take the majority of compute time.
}
\examples{
## Example model run
system.time({
reference_two_state_markov()
})
}
\author{
Howard Thom
}
