% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_markov.R
\name{sample_markov}
\alias{sample_markov}
\title{Sample a Markov Model}
\usage{
sample_markov(markov_model = NULL, type = "base")
}
\arguments{
\item{markov_model}{A list of functions that define a markov model across multiple interventions. See \code{example_two_state_markov}
for the correct format.}

\item{type}{A character string specifying the approach to use to simulate the model. Currently implemented
approaches are "base" with "base" as the default.}
}
\value{
A single sample of a model encoded in the \code{SpeedyMarkov} format (see \code{example_two_state_markov} for details).
}
\description{
This model agnostic function samples a single markov model specification. It wraps multiple approaches
that may offer various advantages and disadvantages.
}
\examples{

sample <- sample_markov(example_two_state_markov())
  
sample
}
