% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_markov.R
\name{simulate_markov}
\alias{simulate_markov}
\title{Simulate a Markov Model Sample}
\usage{
simulate_markov(
  markov_sample = NULL,
  duration = NULL,
  discount = 1.035,
  type = "base"
)
}
\arguments{
\item{markov_sample}{A single row dataframe or a list with no default. See \code{sample_markov}
for the correct data format.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{type}{A character string specifying the approach to use to simulate the model. Currently implemented
approaches are "base" with "base" as the default.}
}
\value{
A list containing total costs and total QALYs as matrices across states
and the duration of the model
}
\description{
This model agnostic function runs a single markov model for the specified duration. It wraps multiple approaches
that may offer various advantages and disadvantages.
}
\examples{

sample <- sample_markov(example_two_state_markov())
  
simulate_markov(sample[1, ], duration = 10)
  
}
