% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_markov.R
\name{simulate_markov_base}
\alias{simulate_markov_base}
\title{Simulate a Markov Model Sample using Base R}
\usage{
simulate_markov_base(
  transition = NULL,
  cohort = NULL,
  state_cost = NULL,
  intervention_cost = NULL,
  qalys = NULL,
  duration = NULL,
  discount = 1.035,
  sim = NULL
)
}
\arguments{
\item{transition}{A transition matrix, see \code{example_two_state_markov} for an example of setting this up.}

\item{cohort}{A list containing the initial state for each intervention,
see \code{example_two_state_markov} for an example of setting this up.}

\item{state_cost}{A list of state costs for each intervention,
see \code{example_two_state_markov} for an example of setting this up.}

\item{intervention_cost}{A vector of intervention costs, see \code{example_two_state_markov} for an example of setting this up.}

\item{qalys}{A list of QALYs for each intervention, see \code{example_two_state_markov} for an example of setting this up.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{sim}{Matrix with the same number of rows as the duration of the model and the same number of columns as the number of
states in the model. Used to store model simulatons. Rather than setting up for each sample this can be preallocated initially to increase efficiency.}
}
\value{
A list containing total costs and total QALYs as matrices across states
and the duration of the model
}
\description{
This model agnostic function runs a single markov model for the specified duration using a base R implementation.
See \code{example_two_state_markov} for an example of the required input. Alternatively use \code{sample_markov(type = "base")} and
the output from \code{sample_markov}.
}
\examples{
  
 markov_sample <- sample_markov(example_two_state_markov())
 
 simulate_markov_base(
     transition = markov_sample$transition[[1]],
     cohort = markov_sample$cohort[[1]],
     state_cost = markov_sample$state_cost[[1]], 
     intervention_cost = markov_sample$intervention_cost[[1]], 
     qalys = markov_sample$qalys[[1]], 
     duration = 10
 )  
  
  
}
