% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedyMarkov.R
\docType{package}
\name{SpeedyMarkov-package}
\alias{SpeedyMarkov}
\alias{SpeedyMarkov-package}
\title{SpeedyMarkov: Speed up Discrete Markov Model Simulations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Speed up Discrete Markov Model Simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/SpeedyMarkov}
  \item \url{https://github.com/seabbs/SpeedyMarkov}
  \item Report bugs at \url{https://github.com/seabbs/SpeedyMarkov/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

}
\keyword{internal}
