% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_arrange.R
\name{vector_arrange}
\alias{vector_arrange}
\title{Arrange Vectorised Vector Samples}
\usage{
vector_arrange(samples)
}
\arguments{
\item{samples}{A list of vectorised samples}
}
\value{
A list of vectors with each vector representing a single sample.
}
\description{
A convenience function used to arrange vectorised samples into the correct
vector format for several functions used to specify Markov model. See \code{example_two_state_markov}
for an example use case.
}
\examples{

vector_samples <- list(rnorm(10, 1, 2), rnorm(10, 10, 2))

vector_arrange(vector_samples)
}
