% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_ce_pipeline.R
\name{markov_ce_pipeline}
\alias{markov_ce_pipeline}
\title{Markov Sampling, Simulation, and Cost Effectiveness Analysis Pipeline}
\usage{
markov_ce_pipeline(
  markov_model = NULL,
  duration = NULL,
  discount = 1.035,
  samples = 1,
  baseline = 1,
  willingness_to_pay_thresold = 20000,
  type = "base"
)
}
\arguments{
\item{markov_model}{A list of functions that define a markov model across multiple interventions. See \code{example_two_state_markov}
for the correct format.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{samples}{Numeric, defaults to 1. The number of markov model samples to use.}

\item{baseline}{Numeric, the intervetion to consider as the baseline for pairwise comparisons.}

\item{willingness_to_pay_thresold}{Numeric, defaulting to 20,000. This is the thresold at which an intervention
may be considered cost effective in the UK.}

\item{type}{A character string specifying the approach to use in the modelling pipeline. Currently implemented
approaches are "base" with "base" as the default.}
}
\value{
A list containing the model samples and simulations and cost effectiveness summary measures.
}
\description{
Markov Sampling, Simulation, and Cost Effectiveness Analysis Pipeline
}
\examples{

markov_ce_pipeline(example_two_state_markov(), duration = 10, samples = 5)
  
}
\seealso{
markov_simulation_pipeline analyse_ce
}
