% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_simulation_pipeline.R
\name{markov_simulation_pipeline}
\alias{markov_simulation_pipeline}
\title{Markov Sampling and Simulation Pipeline}
\usage{
markov_simulation_pipeline(
  markov_model = NULL,
  duration = NULL,
  discount = 1.035,
  samples = 1,
  type = "base",
  debug = FALSE
)
}
\arguments{
\item{markov_model}{A list of functions that define a markov model across multiple interventions. See \code{example_two_state_markov}
for the correct format.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discount}{Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.}

\item{samples}{Numeric, defaults to 1. The number of markov model samples to use.}

\item{type}{A character string specifying the approach to use in the modelling pipeline. Currently implemented
approaches are "base" with "base" as the default.}

\item{debug}{Logical, defaults to \code{FALSE}. Turns on all debug checks - this may impact runtimes.}
}
\value{
A list containing the model samples and simulations.
}
\description{
Markov Sampling and Simulation Pipeline
}
\examples{


markov_simulation_pipeline(example_two_state_markov(), duration = 10, samples = 2)
  
}
\seealso{
sample_markov simulate_markov
}
