% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_markov.R
\name{simulate_markov}
\alias{simulate_markov}
\title{Simulate a Markov Model Sample}
\usage{
simulate_markov(
  markov_sample = NULL,
  duration = NULL,
  discounting = NULL,
  type = "base",
  debug = FALSE,
  input_is_list = NULL,
  sim = NULL
)
}
\arguments{
\item{markov_sample}{A single row dataframe or a list with no default. See \code{sample_markov}
for the correct data format.}

\item{duration}{Numeric, how many long to run the model for.}

\item{discounting}{Numeric vector, the discount that should be applied to the costs and QALYs for each time period.
This must be the same legnth as \code{duration}.}

\item{type}{A character string specifying the approach to use to simulate the model. Currently implemented
approaches are "base" with "base" as the default.}

\item{debug}{Logical, defaults to \code{FALSE}. Turns on all debug checks - this may impact runtimes.}

\item{input_is_list}{Logical defaults to NULL. What type of input is \code{markov_sample}? A list or a dataframe.
If not given then the input type will be checked and converted to a list as required.}

\item{sim}{Matrix with the same number of rows as the duration of the model and the same number of columns as the number of
states in the model. Used to store model simulatons. Rather than setting up for each sample this can be preallocated initially to increase efficiency.}
}
\value{
A list containing total costs and total QALYs as matrices across states
and the duration of the model
}
\description{
This model agnostic function runs a single markov model for the specified duration. It wraps multiple approaches
that may offer various advantages and disadvantages.
}
\examples{

sample <- sample_markov(example_two_state_markov())
  
simulate_markov(sample[1, ], duration = 10, discounting = SpeedyMarkov::calc_discounting(1.035, 10))
  
}
